pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bybit-trading-analysis"
version = "0.1.0"
description = "Trading analysis tool for Bybit with Google Drive and email reporting"
authors = [
    {name = "Glenn Santoso", email = "glennstevensantoso@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pybit>=2.4.0",
    "pandas>=1.5.0",
    "matplotlib>=3.5.0",
    "python-dotenv>=1.0.0",
    "questionary>=1.10.0",
    "seaborn>=0.12.0", 
    "pydantic>=2.5.3",
    "google-api-python-client>=2.105.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "pre-commit>=3.3.0"
]

[project.urls]
"Homepage" = "https://github.com/glenn-s-santoso/trade-journal-analysis"
"Bug Tracker" = "https://github.com/glenn-s-santoso/trade-journal-analysis/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
